+++ src/main/java/com/fooeating/member/action/ChangePwAction.java
@@ -1,4 +1,4 @@
-package com.fooeating.action;
+package com.fooeating.member.action;
 
 import javax.servlet.http.HttpServletRequest;
 import javax.servlet.http.HttpServletResponse;
@@ -9,13 +9,13 @@ import com.fooeating.commons.ActionForward;
 import com.fooeating.commons.JSForward;
 import com.fooeating.db.PublicDAO;
 
-public class ReviewUpdateAction implements Action {
+public class ChangePwAction implements Action {
 
 	@Override
 	public ActionForward execute(HttpServletRequest request, 
 			HttpServletResponse response) throws Exception {
 		
-		System.out.println(" M: ReviewUpdateAction_execute() 호출 ");
+System.out.println(" M: ChangePwAction_execute() 호출 ");
 		
 		// 세션정보 제어
 		HttpSession session = request.getSession();
@@ -33,15 +33,15 @@ public class ReviewUpdateAction implements Action {
 		request.setCharacterEncoding("UTF-8");
 		
 		// 전달정보 저장(변수)
-		int review_num = Integer.parseInt(request.getParameter("review_num"));
-		String newContent = request.getParameter("newContent");
+		String pw = request.getParameter("pw");
+		String newPw = request.getParameter("newPw");
 		
 		
 		
 		// DAO - 정보수정 메서드 호출 -changePw(id, pw, newPw)
 		PublicDAO dao = new PublicDAO();
 		
-		int result = dao.changeReview(id, review_num, newContent);
+		int result = dao.changePw(id, pw, newPw);
 		
 		System.out.println(" M : 정보 수정 결과 " + result);
 		
@@ -53,10 +53,16 @@ public class ReviewUpdateAction implements Action {
 			return null; // 컨트롤러에서는 이동X
 		}
 		
+		if (result == 0 ) { // 비밀번호 오류
+			// JS이동 처리 객체 동작 호출
+			JSForward.alertAndBack(response, "비밀번호 오류!!");
+			
+			return null; // 컨트롤러에서는 이동X
+		}
 		
-		session.removeAttribute("dto");
+		session.invalidate();
 		// result == 1
-		JSForward.alertAndCloseAndMove(response, "리뷰 수정 완료!","./MyReview.foo");
+		JSForward.alertAndClose(response, "비밀번호 변경 완료!");
 		
 		return null;
 	}
