+++ src/main/java/com/fooeating/member/action/IdCheckAction.java
@@ -1,4 +1,4 @@
-package com.fooeating.action;
+package com.fooeating.member.action;
 
 import java.io.PrintWriter;
