+++ src/main/java/com/fooeating/member/action/MemberJoinAction.java
@@ -1,4 +1,4 @@
-package com.fooeating.action;
+package com.fooeating.member.action;
 
 import java.sql.Timestamp;
