+++ src/main/java/com/fooeating/member/action/MemberUpdateProAction.java
@@ -1,4 +1,6 @@
-package com.fooeating.action;
+package com.fooeating.member.action;
+
+import java.io.PrintWriter;
 
 import javax.servlet.http.HttpServletRequest;
 import javax.servlet.http.HttpServletResponse;
@@ -8,14 +10,15 @@ import com.fooeating.commons.Action;
 import com.fooeating.commons.ActionForward;
 import com.fooeating.commons.JSForward;
 import com.fooeating.db.PublicDAO;
+import com.fooeating.db.UserDTO;
 
-public class ChangePwAction implements Action {
+
+
+public class MemberUpdateProAction implements Action {
 
 	@Override
-	public ActionForward execute(HttpServletRequest request, 
-			HttpServletResponse response) throws Exception {
-		
-System.out.println(" M: ChangePwAction_execute() 호출 ");
+	public ActionForward execute(HttpServletRequest request, HttpServletResponse response) throws Exception {
+		System.out.println(" M: MemberUpdateProAction_execute() 호출 ");
 		
 		// 세션정보 제어
 		HttpSession session = request.getSession();
@@ -32,16 +35,18 @@ System.out.println(" M: ChangePwAction_execute() 호출 ");
 		// 한글처리(인코딩)
 		request.setCharacterEncoding("UTF-8");
 		
-		// 전달정보 저장(변수)
-		String pw = request.getParameter("pw");
-		String newPw = request.getParameter("newPw");
+		// 전달정보 저장(DTO)
+		UserDTO dto = new UserDTO();
+		dto.setUser_id(request.getParameter("user_id"));
+		dto.setPw(request.getParameter("pw"));
+		dto.setName(request.getParameter("name"));
+		dto.setEmail(request.getParameter("email"));
+		dto.setPhone(request.getParameter("phone"));
 		
-		
-		
-		// DAO - 정보수정 메서드 호출 -changePw(id, pw, newPw)
+		// DAO - 정보수정 메서드 호출 - memberUpdate(dto)
 		PublicDAO dao = new PublicDAO();
 		
-		int result = dao.changePw(id, pw, newPw);
+		int result = dao.memberUpdate(dto);
 		
 		System.out.println(" M : 정보 수정 결과 " + result);
 		
@@ -60,12 +65,10 @@ System.out.println(" M: ChangePwAction_execute() 호출 ");
 			return null; // 컨트롤러에서는 이동X
 		}
 		
-		session.invalidate();
 		// result == 1
-		JSForward.alertAndClose(response, "비밀번호 변경 완료!");
+		JSForward.alertAndMove(response, "수정 완료","./Main.foo");
 		
 		return null;
 	}
-		
 
 }
