+++ src/main/webapp/index.jsp
@@ -21,9 +21,9 @@
 	
 	// 가상주소 바로가기
 // 	response.sendRedirect("./가상주소.foo");
-// 	response.sendRedirect("./Main.foo");
+	response.sendRedirect("./Main.foo");
 // 	response.sendRedirect("./MyPageMember.foo");
-	response.sendRedirect("./NoticeList.foo");
+// 	response.sendRedirect("./NoticeList.foo");
 // 	response.sendRedirect("./FaqList.foo");
